requirements.txt
streamlit>=1.36
openai>=1.40
python-docx>=1.1
reportlab>=4.2
docx2txt>=0.8
pdfplumber>=0.11
